add max instances

+++ build.rs
@@ -1,6 +1,6 @@
 use std::collections::HashSet;
 use std::fs::{self, File};
-use std::io::{self, Write};
+use std::io::{self, Read, Write};
 use std::path::Path;
 //helper
 fn to_camel_case(s: &str) -> String {
@@ -26,7 +26,6 @@ fn mod_create() -> Result<(), Box<dyn std::error::Error>> {
 
     for module_dir in module_dirs {
         // Tell Cargo to rerun build.rs if any file in the folder changes
-        println!("cargo:rerun-if-changed={}", module_dir.display());
 
         // Collect all `.rs` files except mod.rs
         let mut mods = HashSet::new();
@@ -57,23 +56,26 @@ fn mod_create() -> Result<(), Box<dyn std::error::Error>> {
         // Sort for consistent output
         let mut sorted_mods: Vec<String> = mods.into_iter().collect();
         sorted_mods.sort();
-
-        // Generate mod.rs content
-        let mut mod_file = File::create(module_dir.join("mod.rs"))?;
+        let mut content = String::new();
         for m in sorted_mods {
-            writeln!(mod_file, "mod {};\npub use {}::*;", m, m)?;
+            content.push_str(&format!("mod {};\npub use {}::*;\n", m, m));
+        }
+        if let Ok(existing_content) = fs::read_to_string(module_dir.join("mod.rs")) {
+            if existing_content != content {
+                // Only write if content has changed to avoid unnecessary rebuilds
+                // Generate mod.rs content
+                eprintln!("cargo:rerun-if-changed={}", module_dir.display());
+                let mut mod_file = File::create(module_dir.join("mod.rs"))?;
+                mod_file.write_all(content.as_bytes())?;
+            }
         }
     }
     Ok(())
 }
 fn asset_create() -> Result<(), Box<dyn std::error::Error>> {
     let asset_dir = Path::new("assets");
-    println!("cargo:rerun-if-changed={}", asset_dir.display());
-    let mut mod_file = File::create(Path::new(r"src/client/components/_common/assets.rs"))?;
-    writeln!(
-        mod_file,
-        "use dioxus::prelude::*;\npub struct ASSETS;\nimpl ASSETS {{"
-    )?;
+    let mut str_builder: String = String::new();
+    str_builder.push_str("use dioxus::prelude::*;\npub struct ASSETS;\nimpl ASSETS {");
     let paths = fs::read_dir(asset_dir)?
         .map(|res| res.map(|e| e.path()))
         .collect::<Result<Vec<_>, io::Error>>()?;
@@ -85,12 +87,11 @@ fn asset_create() -> Result<(), Box<dyn std::error::Error>> {
                 match ext.to_string_lossy().as_ref() {
                     "png" | "ico" | "svg" => {
                         if let Some(file_name) = path.file_name() {
-                            writeln!(
-                                mod_file,
-                                "    pub const {}: Asset = asset!(\"assets/{}\");",
+                            str_builder.push_str(&format!(
+                                "    pub const {}: Asset = asset!(\"assets/{}\");\n",
                                 file_name.to_string_lossy().replace('.', "_").to_uppercase(),
                                 file_name.to_string_lossy()
-                            )?;
+                            ));
                         }
                     }
                     _ => {}
@@ -98,14 +99,20 @@ fn asset_create() -> Result<(), Box<dyn std::error::Error>> {
             }
         }
     }
-    writeln!(mod_file, "}}")?;
+    str_builder.push_str("}");
+    if let Ok(existing_content) = fs::read_to_string("src/client/components/_common/assets.rs") {
+        if existing_content != str_builder {
+            let mut mod_file = File::create(Path::new(r"src/client/components/_common/assets.rs"))?;
+            eprintln!("cargo:rerun-if-changed={}", asset_dir.display());
+            mod_file.write_all(str_builder.as_bytes())?;
+        }
+    }
     Ok(())
 }
 fn auto_route(base_path: &str) -> Result<(), Box<dyn std::error::Error>> {
     let views_dir = Path::new(base_path);
-    println!("cargo:rerun-if-changed={}", views_dir.display());
-    let mut mod_file = File::create(views_dir.join("mod.rs"))?;
-    writeln!(mod_file, "use dioxus::prelude::*;")?;
+    let mut str_builder: String = String::new();
+    str_builder.push_str("use dioxus::prelude::*;\n");
     let mut views = HashSet::new();
     let paths = fs::read_dir(views_dir)?
         .map(|res| res.map(|e| e.path()))
@@ -123,12 +130,11 @@ fn auto_route(base_path: &str) -> Result<(), Box<dyn std::error::Error>> {
                 if ext == "rs" {
                     if let Some(file_name) = path.file_stem() {
                         if file_name != "mod" {
-                            writeln!(
-                                mod_file,
+                            str_builder.push_str(&format!(
                                 "mod {};\npub use {}::*;",
                                 file_name.to_string_lossy(),
                                 file_name.to_string_lossy()
-                            )?;
+                            ));
                             views.insert((
                                 relative_file_path_without_ext.to_string_lossy().to_string(),
                                 to_camel_case(&file_name.to_string_lossy().to_string()),
@@ -140,35 +146,38 @@ fn auto_route(base_path: &str) -> Result<(), Box<dyn std::error::Error>> {
         }
         if path.is_dir() {
             if let Some(file_name) = path.file_name() {
-                writeln!(
-                    mod_file,
+                str_builder.push_str(&format!(
                     "mod {};\npub use {}::*;",
                     file_name.to_string_lossy(),
                     file_name.to_string_lossy()
-                )?;
+                ));
             }
         }
     }
-    writeln!(
-        mod_file,
-        "#[derive(Debug, Clone, Routable, PartialEq)]\n#[rustfmt::skip]\npub enum Route {{\n    #[layout(Layout)]"
-    )?;
+    str_builder.push_str("#[derive(Debug, Clone, Routable, PartialEq)]\n#[rustfmt::skip]\npub enum Route {\n    #[layout(Layout)]");
     for page in views {
         if page.0.is_empty() || page.0 == "home" {
-            writeln!(mod_file, "        #[route(\"/\")]\n        {},", page.1)?;
+            str_builder.push_str(&format!("        #[route(\"/\")]\n        {},", page.1));
         } else {
-            writeln!(
-                mod_file,
+            str_builder.push_str(&format!(
                 "        #[route(\"/{}\")]\n        {},",
                 page.0.replace('\\', "/"),
                 page.1
-            )?;
+            ));
         }
     }
-    writeln!(mod_file, "}}")?;
+    str_builder.push_str("}");
+    if let Ok(existing_content) = fs::read_to_string(views_dir.join("mod.rs")) {
+        if existing_content != str_builder {
+            let mut mod_file = File::create(views_dir.join("mod.rs"))?;
+            eprintln!("cargo:rerun-if-changed={}", views_dir.display());
+            mod_file.write_all(str_builder.as_bytes())?;
+        }
+    }
     Ok(())
 }
 fn main() -> io::Result<()> {
+    eprintln!("cargo:rerun-if-changed=build.rs");
     _ = mod_create();
     _ = asset_create();
     _ = auto_route("src/client/views");
